refactor: recipe mapper 함수 분리
