+++ app/src/main/java/com/surivalcoding/composerecipeapp/data/model/Recipe.kt
@@ -1,7 +1,5 @@
 package com.surivalcoding.composerecipeapp.data.model
 
-import com.surivalcoding.composerecipeapp.data.mock.fakeAvatarImage
-import com.surivalcoding.composerecipeapp.data.mock.fakeImageDrawables
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -17,38 +15,4 @@ data class Recipe(
     val authorName: String,
     val ingredients: List<Ingredient>,
     val videoLink: String,
-)
-
-fun Recipe.toSavedRecipe() = SavedRecipe(
-    id = id,
-    thumbnailUrl = thumbnailUrl,
-    cookingMinute = cookingMinute,
-    title = title,
-    authorName = authorName,
-    rating = rating,
-)
-
-fun Recipe.toHomeRecipe() = HomeRecipe(
-    id = id,
-    foodIconUrl = fakeImageDrawables[(id - 1) % 3],
-    cookingMinute = cookingMinute,
-    title = title,
-    rating = rating,
-)
-
-fun Recipe.toNewRecipe() = NewRecipe(
-    id = id,
-    foodIconUrl = fakeImageDrawables[(id + 1) % 3],
-    cookingMinute = cookingMinute,
-    title = title,
-    rating = rating,
-    authorName = authorName,
-    authorAvatarUrl = fakeAvatarImage[(id - 1) % 3],
-)
-
-fun Recipe.toSearchRecipe() = SearchRecipe(
-    thumbnailUrl = thumbnailUrl,
-    title = title,
-    rating = rating,
-    authorName = authorName
 )
